[electron] Test codecs from ffmpeg
Electron's ffmpeg can bundle proprietary codecs, which we want to avoid
having.

This commit adds a native module that will be built and run on the
client where the `@theia/electron` package will be installed.

The program will link at runtime against Electron's ffmpeg shared
library and scan the available codecs, failing if a proprietary one is
found.

Documentation about ffmpeg's API can be found at:
https://www.ffmpeg.org/doxygen/3.4/index.html

The test is ran on post-install and will fail the process if proprietary
codecs are found. Since Yarn suppresses the sub-packages output, I write
and display a log file on post-install, the timestamp should help
identify if the log file was written during the last build or not.

Note: On windows, node-gyp's path resolution is broken, so I had to
create a js file to correctly resolve node-gyp's cli.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>